Drop dependency on pandoc_ast
Closes #67

Instead of using pandoc_ast to ingest the JSON representation of
pandoc's AST, we are now using pandoc to convert the md to html, then
using scraper to extract the parts the we need, and html2md to return
render anything back to md if needed.

This is intended as a stopgap.